feat: 알림 설정 뎊스 수정
